Split functions.py into 5 files
